requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
